[delegationManager] catch Unix_error (fix coq-community/vscoq#200)
This commit exposes a little bug in the current API for swapping
the implementation of par. In particular the code which is called
is not passed the Stateid.t of the current sentence, so it can't
easily report and error without assuming the execution manager did
set some imperative state holding the current sentence id.